+++ src/test/resources/com/atomist/rug/runtime/js/interop/SimpleCommandHandlerWithNullDefault.ts
@@ -1,4 +1,4 @@
-import {HandleCommand, Instruction, Response, HandlerContext, Plan, ResponseMessage} from '@atomist/rug/operations/Handlers'
+import {HandleCommand, Instruction, Response, HandlerContext, CommandPlan, ResponseMessage} from '@atomist/rug/operations/Handlers'
 import {CommandHandler, Parameter, Tags, Intent} from '@atomist/rug/operations/Decorators'
 
 @CommandHandler("ShowMeTheKitties","Search Youtube for kitty videos and post results to slack")
@@ -7,8 +7,8 @@ class KittieFetcher implements HandleCommand{
   @Parameter({description: "test", pattern: "@any", required: false})
   test: string = null
 
-  handle(ctx: HandlerContext) : Plan {
-    return new Plan().add(new ResponseMessage("body"));
+  handle(ctx: HandlerContext) : CommandPlan {
+    return new CommandPlan().add(new ResponseMessage("body"));
   }
 }