+++ src/test/resources/com/atomist/rug/runtime/js/interop/SimpleCommandWithPresentable.ts
@@ -1,4 +1,4 @@
-import {HandleCommand, MappedParameters, ResponseMessage, Instruction, Response, HandlerContext, Plan} from '@atomist/rug/operations/Handlers'
+import {HandleCommand, MappedParameters, ResponseMessage, Instruction, Response, HandlerContext, CommandPlan} from '@atomist/rug/operations/Handlers'
 import {CommandHandler, Parameter, MappedParameter, Tags, Intent} from '@atomist/rug/operations/Decorators'
 
 @CommandHandler("ShowMeTheKitties","Search Youtube for kitty videos and post results to slack")
@@ -12,13 +12,13 @@ class KittieFetcher implements HandleCommand {
   @MappedParameter(MappedParameters.GITHUB_REPO_OWNER)
   owner: string
 
-  handle(ctx: HandlerContext) : Plan {
+  handle(ctx: HandlerContext) : CommandPlan {
     let pxe = ctx.pathExpressionEngine
 
     if(this.name != "el duderino") {
       throw new Error("This will not stand");
     }
-    let result = new Plan()
+    let result = new CommandPlan()
     let message = new ResponseMessage("KittieFetcher");
     message.addAction({ instruction: {
                  kind: "command",